Move preprocessing task specification from ROI profiles to processing tasks
Preprocessing tasks have required an ROI profile setting in addition to the `--proc preprocess` CLI argument. Although a custom ROI profile may be necessary for several preprocessing tasks such as saturation and denoising, the profile is not necessary for others such as rotation, meaning that the ROI profile may be required only for this setting.

As a simplification, integrate preprocessing settings into the processing tasks CLI argument. `--proc preprocess=rotate` can now be used, for example, to specify a rotation preprocessing task, without requiring a separate ROI profile. Move the `PreProcessKeys` to `config` for consistency, and remove the associated ROI profile argument. Removing this argument also reduces potential confusion with the atlas profile setting of the same name.
